Line translation for line history
If there are staged or worktree changes, then the lines selected for the line history command might not correspond the lines the user actually wants to see. This adds a new `LineTranslator` class and some logic to translate editor lines to `HEAD` file lines to solve this issue by parsing through the worktree and staged diffs.