Consolidate all refs into a single store
This still has the issue of refs of differents types overriding each other. For example, you could have a branch and a tag with the same name. In this case, it's just last-wins for which ref type is store in `RefStore`. I need a way to handle collisions when adding a string.