Improve reference and filename watching
This accounts for the adding/removing of references with slashes, which are implemented as nested folders. When all of the references under a folder are deleted, Git just deletes the folder, which we need to account for with our file system watchers. Things are even trickier when removing references, since we don't know whether the deleted URI is a file or a folder.